Build lexlib to support shared object lexers
In Scintilla 3.7.2, the only way to add an external lexer without
modifying the source code to Scintilla itself is to use an external
lexer implemented as a shared library.  While his can be done from
scratch by implementing the necessary functions in a shared library, the
easiest way to implement an external lexer is using the helper classes
in Scintilla's lexlib library.