Stub most lexer functions; fix build dependency on plugin
The custom command needs to be on the plugin target to copy the plugin
to the directories of the other targets.  This corrects the issue where
you have a failing unit test and then update the implementation to pass
the test.  Whenever the plugin is built, it is copied to the target
directories that need the plugin.  With the custom command on the
dependent targets, rebuilding the plugin didn't copy the plugin to the
dependent target directories.