Go back to POST_BUILD event mechanism
The custom command has to be placed on the dependency and apparently it
also depends on the order in which targets are defined (which seems
surprising), otherwise a build from scratch fails.  add_custom_command
will set up the necessary dependencies, so the original call to
add_dependencies is redundant.

This still requires a manually forced build of the test or tool to get
an updated lexer plugin copied.